Adapt annotator auth into local oauth provider
