+++ development.ini
@@ -12,7 +12,7 @@ horus.require_activation: False
 horus.login_redirect: /
 horus.logout_redirect: /
 
-multiauth.policies: h.api.token h.models
+multiauth.policies: h.auth.local h.api
 
 mail.default_sender: "Annotation Daemon" <no-reply@localhost>
