--- h/exceptions.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from horus.exceptions import AuthenticationFailure
-
-__all__ = ['AuthenticationFailure']
