+++ production.ini
@@ -1,13 +1,6 @@
 [app:main]
 use: egg:h
 
-# Authentication
-#
-# If authentications must be shared across servers or persisted across server
-# reloads then this value should be set to a random 64-character string.
-# Keep this value secure!
-#auth.secret=
-
 # API configuration
 #
 # Customize the key or leave it as the default. If the key is present without
@@ -34,11 +27,18 @@ use: egg:h
 api.endpoint: /api
 api.key: 00000000-0000-0000-0000-000000000000
 
+# Authentication settings
+# Add subsections for each provider included with the multiauth setting below.
+# The local annotator API is handled by the api section above.
+#auth.local.authorization: /oauth/authorize
+#auth.local.token: /oauth/token
+
 # ElasticSearch configuration
 #es.host: http://localhost:9200
 #es.index: annotator
 
 # User and group framework settings -- see horus documentation
+# Used by the local authentication provider
 horus.login_redirect: /
 horus.logout_redirect: /
 #horus.activate_redirect: /
@@ -48,7 +48,7 @@ horus.logout_redirect: /
 #horus.require_activation: True
 
 # Authorization settings -- see pyramid_multiauth documentation
-multiauth.policies: h.api h.models
+multiauth.policies: h.auth.local h.api
 
 # Mail server configuration -- see the pyramid_mailer documentation
 mail.default_sender: "Annotation Daemon" <no-reply@localhost>
